Consertando a classe UsuarioController

+++ GerenciamentoDeEstoqueFarmaciaBack/src/main/java/com/LucasVicentee/GerenciamentoDeEstoque/repositories/UsuarioRepository.java
@@ -4,7 +4,7 @@ import com.LucasVicentee.GerenciamentoDeEstoque.entities.Usuario;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
 
     Optional<Usuario> findByUsuario(String usuario);
     Optional<Usuario> findByEmail(String email);
